chg: separate sdk clients for Plasma and POS interaction
