+++ src/common/SDKClient.ts
@@ -1,0 +1,101 @@
+import BN from 'bn.js'
+
+import Web3Client from '../common/Web3Client'
+import { address, SendOptions } from '../types/Common'
+import ContractsBase from '../common/ContractsBase'
+
+export default class Matic extends ContractsBase {
+  public web3Client: Web3Client
+
+  constructor(options: any = {}) {
+    const web3Client = new Web3Client(
+      options.parentProvider,
+      options.maticProvider,
+      options.parentDefaultOptions || {},
+      options.maticDefaultOptions || {}
+    )
+    super(web3Client)
+    this.web3Client = web3Client
+  }
+
+  setWallet(_wallet) {
+    this.web3Client.wallet = _wallet
+  }
+
+  async balanceOfERC20(userAddress: address, token: address, options?: SendOptions) {
+    if (options && (!token || !userAddress)) {
+      throw new Error('token address or user address is missing')
+    }
+
+    const balance = this.getERC20TokenContract(token, options.parent)
+      .methods.balanceOf(userAddress)
+      .call()
+    return balance
+  }
+
+  async balanceOfERC721(userAddress: address, token: address, options?: SendOptions) {
+    if (options && (!token || !userAddress)) {
+      throw new Error('token address or user address is missing')
+    }
+
+    const balance = this.getERC721TokenContract(token, options.parent)
+      .methods.balanceOf(userAddress)
+      .call()
+    return balance
+  }
+
+  async tokenOfOwnerByIndexERC721(userAddress: address, token: address, index: number, options?: SendOptions) {
+    if (options && (!token || !userAddress)) {
+      throw new Error('token address or user address is missing')
+    }
+
+    const tokenID = this.getERC721TokenContract(token, options.parent)
+      .methods.tokenOfOwnerByIndex(userAddress, index)
+      .call()
+    return tokenID
+  }
+
+  async transferERC20Tokens(token: address, to: address, amount: BN | string, options?: SendOptions) {
+    if (options && (!options.from || !amount || !token || !to)) {
+      throw new Error('options.from, to, token or amount is missing')
+    }
+
+    const txObject = this.getERC20TokenContract(token, options.parent).methods.transfer(to, this.encode(amount))
+
+    const _options = await this._fillOptions(
+      options,
+      txObject,
+      options.parent ? this.web3Client.getParentWeb3() : this.web3Client.getMaticWeb3()
+    )
+
+    if (options.encodeAbi) {
+      _options.data = txObject.encodeABI()
+      _options.to = token
+      return _options
+    }
+
+    return this.web3Client.send(txObject, _options)
+  }
+
+  async transferERC721Tokens(token: address, to: address, tokenId: string, options?: SendOptions) {
+    if (options && (!options.from || !tokenId || !token || !to)) {
+      throw new Error('options.from, to, token or tokenId is missing')
+    }
+
+    const txObject = this.getERC721TokenContract(token, options.parent).methods.transferFrom(options.from, to, tokenId)
+
+    const _options = await this._fillOptions(
+      options,
+      txObject,
+      options.parent ? this.web3Client.getParentWeb3() : this.web3Client.getMaticWeb3()
+    )
+
+    if (options.encodeAbi) {
+      _options.data = txObject.encodeABI()
+      _options.to = token
+      return _options
+    }
+
+    return this.web3Client.send(txObject, _options)
+  }
+}
